pyproject.toml
[tool.poetry]
name = "marcov_tweet"
version = "0.1.0"
description = ""
authors = ["odmishien <tmtm.yagi.007@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
janome = "^0.3.10"
python-twitter = "^3.5"
python-dotenv = "^0.11.0"
markovify = "^0.8.0"
mecab-python3 = "^0.996.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
